IotDet-Backend/requirements.txt
pip==22.3.1
setuptools==65.7.0
wheel==0.38.4
PyMySQL==1.0.2
uvicorn==0.20.0
fastapi==0.89.1
SQLAlchemy==1.4.46
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.5
requests==2.28.2
pytest==7.2.0
httpx==0.23.3
cryptography==39.0.0
pydantic~=1.10.4
pandas==2.0.0
numpy~=1.25.0
matplotlib~=3.7.1
seaborn~=0.12.2
scikit-learn~=1.2.2
keras~=2.13.1rc0
imbalanced-learn~=0.11.0